# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Course API",
        "description": "API documentation for authentication, courses, interests, favorites, students, and teachers.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Application server"
        }
    ],
    "paths": {
...
        "/api/v1/courses/{course}/join": {
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Join a course or start checkout flow",
                "operationId": "750b552abff2dc80d12e1800b7ae6fa3",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Joined course",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Joined course successfully.",
                                    "inscription": {
                                        "id": 1,
                                        "user_id": 10,
                                        "cours_id": 3,
                                        "group_id": 2,
                                        "payment_status": "paid"
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Already enrolled or checkout required",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Please complete payment to join the course.",
                                    "session_id": "cs_test_a1b2c3",
                                    "checkout_url": "https://checkout.stripe.com/c/pay/cs_test_a1b2c3",
                                    "inscription": {
                                        "id": 1,
                                        "user_id": 10,
                                        "cours_id": 3,
                                        "group_id": null,
                                        "payment_status": "unpaid"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "This course has no payable amount.",
                                    "errors": {
                                        "course": [
                                            "This course has no payable amount."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/courses/{course}/checkout": {
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Create checkout session for course payment",
                "operationId": "e70a1b4c7b63d1676bf3737e2d33c829",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checkout session created or already paid",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Checkout session created successfully.",
                                    "session_id": "cs_test_a1b2c3",
                                    "checkout_url": "https://checkout.stripe.com/c/pay/cs_test_a1b2c3",
                                    "inscription": {
                                        "id": 1,
                                        "user_id": 10,
                                        "cours_id": 3,
                                        "group_id": null,
                                        "payment_status": "unpaid"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "This course has no payable amount.",
                                    "errors": {
                                        "course": [
                                            "This course has no payable amount."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/payments/success": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Confirm successful payment by session id",
                "operationId": "08d7534fc40bddb086841bca6bf91b63",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment confirmed",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Payment confirmed and inscription updated.",
                                    "inscription": {
                                        "id": 1,
                                        "user_id": 10,
                                        "cours_id": 3,
                                        "group_id": null,
                                        "payment_status": "paid"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "session_id": [
                                            "The session id field is required."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/courses/{course}/leave": {
            "delete": {
                "tags": [
                    "Students"
                ],
                "summary": "Leave a course",
                "operationId": "1d9f9867a08542bf9e243beba84de651",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Left course or not enrolled",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Left course successfully."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "No query results for model [App\\Models\\Course] 999"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}